refactor(UserService): 인터페이스를 통해 트랜잭션 서비스 추상화
